apps/api/requirements.txt
flask
flask-cors
langchain
python-dotenv